% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ai-ctrl-block.R
\name{ai_ctrl_block}
\alias{ai_ctrl_block}
\alias{ai_ctrl_ui}
\alias{ai_ctrl_server}
\title{AI-powered control block plugin}
\usage{
ai_ctrl_block()

ai_ctrl_ui(id, x)

ai_ctrl_server(id, x, vars, data, eval)
}
\arguments{
\item{id}{Namespace ID}

\item{x}{Block object}

\item{vars}{Reactive state values (pre-filtered to externally controllable vars)}

\item{data}{Input data as list of reactive values}

\item{eval}{Reactive that evaluates block expression against input data}
}
\value{
A ctrl_block plugin object
}
\description{
Replaces the default ctrl_block with an AI chat interface. Users can
describe what they want in natural language and the LLM will configure
the block parameters.
}
\examples{
\dontrun{
serve(
  new_board(new_dataset_block("iris")),
  plugins = custom_plugins(ai_ctrl_block())
)
}

}
