% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-server.R
\name{extract_result}
\alias{extract_result}
\title{Retrieve result}
\usage{
extract_result(x, client)
}
\arguments{
\item{x}{An \code{llm_block_proxy} object used for dispatch}

\item{client}{A \code{ellmer::chat()} object}
}
\description{
From the \code{ellmer::chat()} object, passed as \code{client}, retrieve what is
considered the result of a conversation (given a block of type \code{x}).
}
\keyword{internal}
