% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-llm.R
\name{format_data_preview}
\alias{format_data_preview}
\alias{format_data_preview.data.frame}
\alias{format_data_preview.dm}
\alias{format_data_preview.default}
\title{Format data preview for a single object}
\usage{
format_data_preview(x, ...)

\method{format_data_preview}{data.frame}(x, ...)

\method{format_data_preview}{dm}(x, ...)

\method{format_data_preview}{default}(x, ...)
}
\arguments{
\item{x}{Object to preview}

\item{...}{Additional arguments}
}
\value{
Character string
}
\description{
S3 generic that produces a text summary of an object for LLM consumption.
Packages can define methods for custom types.
}
