% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-blocks.R
\name{get_block_ctor}
\alias{get_block_ctor}
\title{Get block constructor by name}
\usage{
get_block_ctor(block_name)
}
\arguments{
\item{block_name}{Name like "filter_block" or "new_filter_block"}
}
\value{
The constructor function
}
\description{
Get block constructor by name
}
