% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block-signature.R
\name{get_block_signature}
\alias{get_block_signature}
\title{Get block signature information}
\usage{
get_block_signature(block_ctor, name = NULL)
}
\arguments{
\item{block_ctor}{Block constructor function (e.g., new_summarize_block)}
}
\value{
List with:
\itemize{
\item name: Constructor function name
\item params: Named list of parameters with their default values
\item param_names: Character vector of parameter names (excluding ...)
}
}
\description{
Extracts parameter information from a block constructor function.
}
\examples{
\dontrun{
sig <- get_block_signature(blockr.dplyr::new_summarize_block)
sig$param_names
# [1] "summaries" "by"
}

}
