% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-blocks.R
\name{get_dplyr_block_info}
\alias{get_dplyr_block_info}
\title{Get descriptions of available dplyr blocks}
\usage{
get_dplyr_block_info()
}
\value{
Named list with block info (name, description, ctor)
}
\description{
Returns a list of block information suitable for LLM planning.
}
