% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block-signature.R
\name{get_summarize_functions}
\alias{get_summarize_functions}
\title{Get available aggregation functions for summarize block}
\usage{
get_summarize_functions()
}
\value{
Character vector of function names
}
\description{
Returns a list of common aggregation functions that can be used
in the summarize block.
}
