% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm-ui.R
\name{image_upload_js}
\alias{image_upload_js}
\title{JavaScript for image upload functionality}
\usage{
image_upload_js(id)
}
\arguments{
\item{id}{Module namespace id}
}
\description{
JavaScript for image upload functionality
}
