% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-eval.R
\name{inherits_base_class}
\alias{inherits_base_class}
\alias{result_ptype}
\alias{check_result}
\alias{check_result.default}
\alias{check_result.ggplot}
\title{Tools for evaluating and checking code}
\usage{
inherits_base_class(x, ptype)

result_ptype(x)

check_result(x, ptype, check_fun = inherits_base_class)

\method{check_result}{default}(x, ptype, check_fun = inherits_base_class)

\method{check_result}{ggplot}(x, ptype, check_fun = inherits_base_class)
}
\arguments{
\item{x}{Object to be checked}

\item{ptype}{A prototype object}

\item{check_fun}{A function that either returns \code{TRUE} or a string
indicating the type of check failure}
}
\description{
LLM-produced code is evaluated and the result is checked w.r.t. block-
specific expectations.
}
