% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ui.R
\name{llm_block_css}
\alias{llm_block_css}
\alias{llm_block_css.llm_block_proxy}
\alias{llm_block_css.llm_plot_block_proxy}
\title{Block-specific UI}
\usage{
llm_block_css(x)

\method{llm_block_css}{llm_block_proxy}(x)

\method{llm_block_css}{llm_plot_block_proxy}(x)
}
\arguments{
\item{x}{LLM block proxy object}
}
\value{
A string.
}
\description{
UI customization at block level.
}
