% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm-server-det.R
\name{llm_block_ui_det}
\alias{llm_block_ui_det}
\title{UI for deterministic LLM blocks}
\usage{
llm_block_ui_det(x)
}
\arguments{
\item{x}{Block proxy}
}
\value{
UI function
}
\description{
UI for deterministic LLM blocks
}
\keyword{internal}
