% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{make_metadata}
\alias{make_metadata}
\title{Generate metadata from reactive datasets}
\usage{
make_metadata(
  datasets,
  extract_codelist_vars = c("-.*DTC$", "-STUDYID", "-USUBJID", "-DOMAIN", "-SUBJID",
    "-SITEID", "-COUNTRY", "-.*ID$", "-.*NAM$", "PARAM", "LBTEST", "-.*DAT$",
    "-.*DAT_RAW$", "-.*DAT_INT$", "-.*ORRES$", "-.*TIM$", "-LBORESS.*$"),
  max_unique_values = 80,
  tips = ""
)
}
\arguments{
\item{datasets}{A list of named datasets}

\item{extract_codelist_vars}{A character vector of patterns to include/exclude
when extracting codelist variables}

\item{max_unique_values}{The maximum number of unique values to extract (default: 900)}

\item{tips}{Additional tips to include in the context}
}
\value{
A list containing metadata for all datasets
}
\description{
This function generates metadata for a collection of named datasets.
}
