% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block-flxtbl.R, R/block-gt.R,
%   R/block-gtsummary.R, R/block-plot.R, R/block-table-insights.R,
%   R/block-transform.R, R/llm-block.R, R/llm-server.R, R/llm-ui.R
\name{new_llm_flxtbl_block}
\alias{new_llm_flxtbl_block}
\alias{new_llm_gt_block}
\alias{new_llm_gtsummary_block}
\alias{new_llm_plot_block}
\alias{new_llm_table_insights_block}
\alias{new_llm_transform_block}
\alias{new_llm_block}
\alias{result_ptype}
\alias{llm_block_server}
\alias{llm_block_server.llm_block_proxy}
\alias{llm_block_ui}
\alias{llm_block_ui.llm_block_proxy}
\title{LLM block constructor}
\usage{
new_llm_flxtbl_block(...)

new_llm_gt_block(...)

new_llm_gtsummary_block(...)

new_llm_plot_block(...)

new_llm_table_insights_block(
  question = "Analyze key findings, treatment effects, and statistical significance.",
  ...
)

new_llm_transform_block(...)

new_llm_block(
  class,
  question = "",
  code = "",
  explanation = character(),
  max_retries = blockr_option("max_retries", 3L),
  enable_image_upload = FALSE,
  ctor = sys.parent(),
  ...
)

result_ptype(x)

llm_block_server(x)

\method{llm_block_server}{llm_block_proxy}(x)

llm_block_ui(x)

\method{llm_block_ui}{llm_block_proxy}(x)
}
\arguments{
\item{...}{Forwarded to \code{\link[blockr.core:new_block]{blockr.core::new_block()}}}

\item{question}{LLM prompt}

\item{class}{Block subclass}

\item{code}{Code as generated by the LLM}

\item{explanation}{Explanation for the generated code}

\item{max_retries}{Number of retries in case of erroneous code}

\item{enable_image_upload}{Whether to enable image upload functionality}

\item{ctor}{String-valued constructor name or function/frame number (mostly
for internal use or when defining constructors for virtual classes)}

\item{x}{Proxy LLM block object}
}
\value{
All blocks constructed via \code{new_llm_block()} inherit from
\code{llm_block}.
}
\description{
Basic constructor for blocks that inherit from \code{llm_block}. Intended to be
used for more specific block classes such as LLM-based transform or plot
blocks.`)
}
