% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block-insights.R, R/llm-tools.R
\name{llm_tools.llm_insights_block_proxy}
\alias{llm_tools.llm_insights_block_proxy}
\alias{new_llm_tool}
\alias{is_llm_tool}
\alias{get_tool}
\alias{get_prompt}
\alias{llm_tools}
\alias{llm_tools.llm_block_proxy}
\title{LLM tools}
\usage{
\method{llm_tools}{llm_insights_block_proxy}(x, ...)

new_llm_tool(..., prompt = character())

is_llm_tool(x)

get_tool(x)

get_prompt(x)

llm_tools(x, ...)

\method{llm_tools}{llm_block_proxy}(x, ...)
}
\arguments{
\item{x}{Object}

\item{...}{Passed to \code{\link[ellmer:tool]{ellmer::tool()}}.}

\item{prompt}{A string with additional prompts that will be added to the
system prompt.}
}
\value{
All blocks constructed via \code{new_llm_tool()} inherit from
\code{llm_tool}.
}
\description{
Tools can be made available to LLMs in order to make them more powerful and
in turn enabling them to create more accurate results.
}
