% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter.R
\name{reporter_console}
\alias{reporter_console}
\title{Progress reporter: console}
\usage{
reporter_console()
}
\value{
A reporter list
}
\description{
Prints formatted progress to stdout. Useful for interactive standalone
discovery and debugging.
}
