% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block-insights.R, R/utils-prompt.R
\name{system_prompt.llm_insights_block_proxy}
\alias{system_prompt.llm_insights_block_proxy}
\alias{system_prompt}
\alias{system_prompt.default}
\alias{system_prompt.llm_block_proxy}
\alias{describe_inputs}
\alias{describe_input}
\alias{describe_input.flextable}
\alias{describe_input.default}
\title{LLM system prompts}
\usage{
\method{system_prompt}{llm_insights_block_proxy}(x, datasets, tools, ...)

system_prompt(x, ...)

\method{system_prompt}{default}(x, datasets, tools, ...)

\method{system_prompt}{llm_block_proxy}(x, datasets, tools, ...)

describe_inputs(x)

describe_input(x, ...)

\method{describe_input}{flextable}(x, ...)

\method{describe_input}{default}(x, ...)
}
\arguments{
\item{x}{LLM block proxy object}

\item{datasets}{Data sets from which to extract metadata}

\item{tools}{List of \code{\link[ellmer:tool]{ellmer::tool()}} objects}

\item{...}{Generic consistency}
}
\value{
A string.
}
\description{
System prompts for instructing the LLM to return code using certain packages,
adhering to certain styles, etc. are created per LLM block class and a
corresponding S3 generic is available as \code{system_prompt()}, which is
dispatched on LLM block proxy objects.
}
