% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block-transform-det.R
\name{system_prompt_det.llm_transform_block_det_proxy}
\alias{system_prompt_det.llm_transform_block_det_proxy}
\title{System prompt for deterministic transform blocks}
\usage{
\method{system_prompt_det}{llm_transform_block_det_proxy}(x, datasets)
}
\arguments{
\item{x}{Block proxy}

\item{datasets}{Named list of datasets}
}
\value{
Character string with system prompt
}
\description{
System prompt for deterministic transform blocks
}
\keyword{internal}
