% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{validate_block_result}
\alias{validate_block_result}
\title{Validate block result based on block type}
\usage{
validate_block_result(obj, block_proxy)
}
\arguments{
\item{obj}{Result object to validate}

\item{block_proxy}{LLM block proxy object to determine validation type}
}
\value{
List with 'valid' (logical) and 'message' (character)
}
\description{
Validate block result based on block type
}
