% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{validate_ggplot_object}
\alias{validate_ggplot_object}
\title{Validate ggplot objects are properly constructed}
\usage{
validate_ggplot_object(obj)
}
\arguments{
\item{obj}{Object to validate}
}
\value{
List with 'valid' (logical) and 'message' (character)
}
\description{
Validate ggplot objects are properly constructed
}
