% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{blockr.ai-package}
\alias{blockr.ai}
\alias{blockr.ai-package}
\title{blockr.ai: TODO}
\description{
TODO
}
\author{
\strong{Maintainer}: Nicolas Bennett \email{nicolas@cynkra.com}

Authors:
\itemize{
  \item David Granjon \email{david@cynkra.com}
  \item Christoph Sax \email{christoph@cynkra.com}
}

}
\keyword{internal}
