% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-prompt.R
\name{system_prompt}
\alias{system_prompt}
\alias{system_prompt.default}
\alias{system_prompt.llm_block_proxy}
\alias{system_prompt.llm_transform_block_proxy}
\alias{system_prompt.llm_plot_block_proxy}
\alias{system_prompt.llm_gt_block_proxy}
\title{LLM system prompts}
\usage{
system_prompt(x, ...)

\method{system_prompt}{default}(x, ...)

\method{system_prompt}{llm_block_proxy}(x, datasets, ...)

\method{system_prompt}{llm_transform_block_proxy}(x, datasets, ...)

\method{system_prompt}{llm_plot_block_proxy}(x, datasets, ...)

\method{system_prompt}{llm_gt_block_proxy}(x, datasets, ...)
}
\arguments{
\item{x}{LLM block proxy object}

\item{...}{Generic consistency}

\item{datasets}{Data sets from which to extract metadata}
}
\value{
A string.
}
\description{
System prompts for instructing the LLM to return code using certain packages,
adhering to certain styles, etc. are created per LLM block class and a
corresponding S3 generic is available as \code{system_prompt()}, which is
dispatched on LLM block proxy objects.
}
